Flipping selection coefficients
Changed the way selection coefficients are flipped in the parents' generation. Changes to selection coefficients made in a slim script are not retained within .trees output files. Therefore, adding another slim script that's run after neutral mutations get added. This script's sole job is to optionally flip selection coefficients randomly and then save the output as a .txt file. This .txt file can the be loaded in to the Experiment slim script.