History sim new feature
History_sim has a new paramter: "additive_limit"

This is the generation up to which an additive model for fitness is to be used (only relevant when use_dominance == 1).

The idea is that this allows us to run an additive model (which is ~ an order of magnitude faster) for most of the generations, begore switching over to dominance closer to equilibrium.